requirements/base.txt
asgiref==3.5.2
Django==4.1.4
django-environ==0.9.0
django-phonenumber-field==7.0.1
django-timezone-field==5.0
phonenumbers==8.13.2
pytz==2022.6
sqlparse==0.4.3
